Create client.py
Preliminary OS enumeration script, i'm building it like a malware. So a C2 mechanism will be incoming, either through XMPP for whichi already have secured infrastructure or something else.  Just posting this for now, as a sort of 'phase 1 complete' i will be writing a managing module as well so it will be:

Client -> C2 -> Managing Module -> ML